+++ include/sway/tree/layout.h
@@ -3,6 +3,7 @@
 #include <wlr/types/wlr_output_layout.h>
 #include <wlr/render/wlr_texture.h>
 #include "sway/tree/container.h"
+#include "sway/tree/root.h"
 #include "config.h"
 
 enum movement_direction {
@@ -23,28 +24,6 @@ enum resize_edge {
 };
 
 struct sway_container;
-
-struct sway_root {
-	struct wlr_output_layout *output_layout;
-
-	struct wl_listener output_layout_change;
-#ifdef HAVE_XWAYLAND
-	struct wl_list xwayland_unmanaged; // sway_xwayland_unmanaged::link
-#endif
-	struct wl_list drag_icons; // sway_drag_icon::link
-
-	struct wlr_texture *debug_tree;
-
-	struct wl_list outputs; // sway_output::link
-
-	list_t *scratchpad; // struct sway_container
-
-	struct {
-		struct wl_signal new_container;
-	} events;
-};
-
-void layout_init(void);
 
 void container_add_child(struct sway_container *parent,
 		struct sway_container *child);
